requirements.txt
qdrant-client
numpy
pandas
spacy
sentence-transformer
torch
tiktoken
bm25